Allow negative values in duration_ms
The spotify API doesn't make any guarantees, if the returned integers
are signed or unsigned. It turns out, that for the `resume_position_ms`
value in the `ResumePointObject` used for `Episode`s can be negative.
I've only seen the value `-1000`, but don't know when this occurs.

This patch addresses this, by adding a parser for `i64` values that is
called if the attempt to parse `u64` fails.