Remove unnecessary mutable references in clients
OAuthClient::request_token takes self as &mut, yet all of its
implementors are designed for an async environment, where inner
mutability for their fields is required through Arc<Mutex<_>>, as is the
case for the token fields, thus self doesn't actually have to be a
mutable reference. By making the reference immutable, it is much more
comfortable to use the clients in application code, since there's no need to
wrap them e.g. a Mutex or an RwLock.